Fix responsiveness below 940px (not tested)
Below a width of 940px, the direct section-children of #page didn't get a width smaller than 940px. This caused the elements inside to exceed the screen horizontally. Adding a width of 100% fixes this. width: auto; or something similar would do the trick, too. This is not tested thoroughly, only experimented with in the developer console of Chrome. I can't imagine why it won't work elsewhere, though.